+++ crates/module-system/sov-modules-api/src/hooks.rs
@@ -72,7 +72,6 @@ pub trait ApplyBatchHooks<Da: DaSpec> {
     fn end_batch_hook(
         &self,
         _result: &Self::BatchResult,
-        _sender: &Da::Address,
         _state_checkpoint: &mut StateCheckpoint<Self::Spec>,
     ) {
     }