+++ docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+- #1121 changes the `BatchReceipt`'s `inner` default field and adds the `sequencer_da_address` to it. In particular:
+    - Add a new `BatchSequencerReceipt` struct which contains the sequencer's `da_address` and `BatchSequencerOutcome`.
+    - Change the `BatchResult` type from `BatchSequencerOutcome` to `BatchSequencerReceipt`.
+    - Remove the sender's `Da::Address` field from the `end_batch_hook` because it is now accessible from `BatchResult`
+    - Minor stylistic improvements in the `StfBlueprint`.
+- #1120 Improve error handling in SequencerRegistry.
 - #1129 changes `StateVec::iter` to return `Result`s to account for state reading errors. The new method `StateVec::collect_infallible` can be used in tests and for infallible state accessors.
 - #1108 improve error handling in `ProverIncentives` module.
 - #1104 Add `deposit` method to prover incentives module.